[refactor] prompt_util.py を追加し digit_only_key_bindings を切り出し
prompt_toolkit 用のキーバインドヘルパを time_entry_command から分離し、
今後他コマンドでも再利用しやすい形にする。
