+++ src/redi/cli/prompt_util.py
@@ -1,0 +1,13 @@
+from prompt_toolkit.key_binding import KeyBindings
+from prompt_toolkit.keys import Keys
+
+
+def digit_only_key_bindings() -> KeyBindings:
+    kb = KeyBindings()
+
+    @kb.add(Keys.Any)
+    def _(event) -> None:
+        if event.data.isdigit():
+            event.current_buffer.insert_text(event.data)
+
+    return kb
